backend/flask_web_app/requirements.txt
attrs==22.2.0
autopep8==2.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
mysqlclient==2.1.1
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.5.0
Werkzeug==2.2.3
